Homework8
Переделал архитектуру приложения. Исправил проблему с бесконечной прокруткой списка
