+++ app/src/main/java/pt/amn/projectacademy/MainActivity.kt
@@ -2,7 +2,9 @@ package pt.amn.projectacademy
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import pt.amn.projectacademy.models.Movie
+import pt.amn.projectacademy.domain.models.Movie
+import pt.amn.projectacademy.presentation.FragmentMoviesDetails
+import pt.amn.projectacademy.presentation.FragmentMoviesList
 
 class MainActivity : AppCompatActivity(), FragmentMoviesList.MoviesListFragmentClicks,
             FragmentMoviesDetails.MovieDetailsFragmentClicks {
