--- app/src/main/java/pt/amn/projectacademy/MainApplication.kt
@@ -1,19 +1,0 @@
-package pt.amn.projectacademy
-
-import android.app.Application
-import android.content.Context
-
-class MainApplication : Application() {
-
-    init {
-        instance = this
-    }
-
-    companion object {
-        private var instance: MainApplication? = null
-
-        fun applicationContext() : Context {
-            return instance!!.applicationContext
-        }
-    }
-}
