--- app/src/main/java/pt/amn/projectacademy/data/ActorsResponse.kt
@@ -1,17 +1,0 @@
-package pt.amn.projectacademy.data
-
-import kotlinx.serialization.*
-
-@Serializable
-data class ActorsListResponse (
-    val cast: List<Cast>
-)
-
-@Serializable
-data class Cast (
-    val id: Int,
-    val name: String,
-
-    @SerialName("profile_path")
-    val profilePath: String? = null,
-)
