--- app/src/main/java/pt/amn/projectacademy/data/MoviesResponse.kt
@@ -1,46 +1,0 @@
-package pt.amn.projectacademy.data
-
-import kotlinx.serialization.*
-
-@Serializable
-data class MoviesListResponse (
-    val results: List<MovieResponse>,
-    val page: Int
-)
-
-@Serializable
-data class MovieResponse (
-    val adult: Boolean,
-
-    @SerialName("backdrop_path")
-    val backdropPath: String,
-
-    @SerialName("genre_ids")
-    val genreIDS: List<Int>,
-
-    val id: Int,
-
-    @SerialName("original_language")
-    val originalLanguage: String,
-
-    @SerialName("original_title")
-    val originalTitle: String,
-
-    val overview: String,
-    val popularity: Float,
-
-    @SerialName("poster_path")
-    val posterPath: String,
-
-    @SerialName("release_date")
-    val releaseDate: String,
-
-    val title: String,
-    val video: Boolean,
-
-    @SerialName("vote_average")
-    val voteAverage: Float,
-
-    @SerialName("vote_count")
-    val voteCount: Int
-)
