+++ app/src/main/java/pt/amn/projectacademy/data/retrofit/response/ActorsListResponse.kt
@@ -1,0 +1,9 @@
+package pt.amn.projectacademy.data.retrofit.response
+
+import kotlinx.serialization.Serializable
+import pt.amn.projectacademy.data.models.ActorDataModel
+
+@Serializable
+data class ActorsListResponse (
+    val cast: List<ActorDataModel>
+)
