+++ app/src/main/java/pt/amn/projectacademy/data/retrofit/response/GenresListResponse.kt
@@ -1,0 +1,9 @@
+package pt.amn.projectacademy.data.retrofit.response
+
+import kotlinx.serialization.Serializable
+import pt.amn.projectacademy.data.models.GenreDataModel
+
+@Serializable
+data class GenresListResponse (
+    val genres: List<GenreDataModel>
+)
