+++ app/src/main/java/pt/amn/projectacademy/domain/models/Actor.kt
@@ -1,0 +1,11 @@
+package pt.amn.projectacademy.domain.domain
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Actor(
+    val id: Int,
+    val name: String,
+    val picture: String
+) : Parcelable
