--- app/src/main/java/pt/amn/projectacademy/models/Actor.kt
@@ -1,11 +1,0 @@
-package pt.amn.projectacademy.models
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Actor(
-    val id: Int,
-    val name: String,
-    val picture: String
-) : Parcelable
