--- app/src/main/java/pt/amn/projectacademy/models/Genre.kt
@@ -1,7 +1,0 @@
-package pt.amn.projectacademy.models
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Genre(val id: Int, val name: String) : Parcelable
