+++ app/src/main/java/pt/amn/projectacademy/presentation/FragmentMoviesDetails.kt
@@ -1,0 +1,104 @@
+package pt.amn.projectacademy.presentation
+
+import android.content.Context
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.viewModels
+import pt.amn.projectacademy.presentation.adapters.ActorsAdapter
+import pt.amn.projectacademy.databinding.FragmentMoviesDetailsBinding
+import pt.amn.projectacademy.loadImage
+import pt.amn.projectacademy.domain.domain.Actor
+import pt.amn.projectacademy.domain.models.Movie
+import pt.amn.projectacademy.utils.BASE_URL_BACKDROP_IMAGE
+import pt.amn.projectacademy.presentation.viewmodels.MovieDetailsViewModel
+import pt.amn.projectacademy.presentation.viewmodels.MovieDetailsViewModelFactory
+
+private const val PARAM_MOVIE = "fragment_movie"
+
+class FragmentMoviesDetails : Fragment() {
+
+    private var _binding : FragmentMoviesDetailsBinding? = null
+    // This property is only valid between onCreateView and onDestroyView
+    private val binding get() = _binding!!
+
+    private var movie: Movie? = null
+    private var listener : MovieDetailsFragmentClicks? = null
+    private val adapter : ActorsAdapter = ActorsAdapter()
+
+    private val viewModel: MovieDetailsViewModel by viewModels {
+        MovieDetailsViewModelFactory(movie?.id)
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        _binding = FragmentMoviesDetailsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.rvActors.adapter = adapter
+
+        binding.tvPath.setOnClickListener {
+            listener?.backClick()
+        }
+
+        movie = requireArguments().getParcelable(PARAM_MOVIE)
+        movie?.run {
+            binding.tvName.text = title
+            binding.tvStoryline.text = overview
+            binding.tvTag.text = getTag()
+            binding.tvReview.text = getReview()
+            binding.tvAge.text = getMinimumAge()
+            binding.ratingBar.rating = getRating()
+            binding.ivBackground.loadImage(binding.root, BASE_URL_BACKDROP_IMAGE + backdrop)
+        }
+
+        viewModel.actorsList.observe(viewLifecycleOwner) { actorList ->
+            // load the list of actors
+            updateData(actorList)
+        }
+
+    }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        if (context is MovieDetailsFragmentClicks) {
+            listener = context
+        }
+    }
+
+    override fun onDetach() {
+        super.onDetach()
+        listener = null
+    }
+
+    override fun onDestroyView() {
+        _binding = null
+        super.onDestroyView()
+    }
+
+    private fun updateData(actorsList: List<Actor>) {
+        adapter.bindActors(actorsList)
+        adapter.notifyDataSetChanged()
+    }
+
+    companion object {
+
+        /** @return A new instance of fragment FragmentMoviesDetails.*/
+        @JvmStatic
+        fun newInstance(movie : Movie) = FragmentMoviesDetails().apply {
+            arguments = Bundle().apply { putParcelable(PARAM_MOVIE, movie) }
+        }
+    }
+
+    interface MovieDetailsFragmentClicks {
+        fun backClick()
+    }
+
+}
