--- app/src/main/java/pt/amn/projectacademy/utilities/Constants.kt
@@ -1,12 +1,0 @@
-package pt.amn.projectacademy.utilities
-
-import pt.amn.projectacademy.BuildConfig
-
-/**
- * Constants used throughout the app.
- */
-const val TMDB_API_KEY = BuildConfig.TMDB_API_KEY
-
-const val BASE_URL_BACKDROP_IMAGE = "https://image.tmdb.org/t/p/w780"
-const val BASE_URL_POSTER_IMAGE = "https://image.tmdb.org/t/p/w300"
-const val BASE_URL_ACTOR_IMAGE = "https://image.tmdb.org/t/p/w185"
