+++ app/src/main/java/pt/amn/projectacademy/utils/Constants.kt
@@ -1,0 +1,12 @@
+package pt.amn.projectacademy.utils
+
+import pt.amn.projectacademy.BuildConfig
+
+/**
+ * Constants used throughout the app.
+ */
+const val TMDB_API_KEY = BuildConfig.TMDB_API_KEY
+
+const val BASE_URL_BACKDROP_IMAGE = "https://image.tmdb.org/t/p/w780"
+const val BASE_URL_POSTER_IMAGE = "https://image.tmdb.org/t/p/w300"
+const val BASE_URL_ACTOR_IMAGE = "https://image.tmdb.org/t/p/w185"
