--- app/src/main/java/pt/amn/projectacademy/viewmodels/MovieDetailsViewModelFactory.kt
@@ -1,14 +1,0 @@
-package pt.amn.projectacademy.viewmodels
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-class MovieDetailsViewModelFactory(private val movieId: Int?) : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
-        MovieDetailsViewModel::class.java -> MovieDetailsViewModel(movieId ?: 0)
-        else -> throw IllegalArgumentException("$modelClass is not registered ViewModel")
-    } as T
-
-}
